refactor: move out the logout function and update its usages
