+++ src/components/auth/LogoutButton.tsx
@@ -1,37 +1,11 @@
 import React from "react";
 import { useRouter } from "next/navigation";
-import axios from "axios";
+import { logout } from "@/lib/services/auth/logout";
 import { auth } from "@/lib/utils/firebase/config";
 import { Button } from "@/components/ui/button";
 import { useToast } from "@/components/ui/use-toast";
 
 type LogoutButtonProps = {};
-
-/**
- * Logs out the user, removes the session token and redirects to the login page.
- * @param router - The router object from Next.js.
- * @param onError - Optional callback function to handle errors.
- * @param avoidRedirect - Optional flag to avoid redirecting to the login page after logout.
- * @throws Error - If an error occurs while logging out.
- */
-export const logout = async (
-  router: ReturnType<typeof useRouter>,
-  onError?: Function,
-  avoidRedirect?: boolean
-) => {
-  try {
-    auth.currentUser && (await auth.signOut());
-    const response = await axios.post("/api/logout");
-
-    if (!avoidRedirect && response.status === 200) {
-      router.push("/login");
-      return;
-    }
-  } catch (e: Error | any) {
-    console.error(e);
-    onError?.(e);
-  }
-};
 
 /**
  * LogoutButton component.
@@ -46,7 +20,7 @@ const LogoutButton: React.FC<LogoutButtonProps> = () => {
       <Button
         variant={"outline"}
         onClick={() =>
-          logout(router, (e: Error | any) => {
+          logout(auth, router, (e: Error | any) => {
             const description = e?.message || "Please try again.";
             displayToast({
               title: "An error has occurred while logging out!",
